requirements.txt
pandas==2.2.3
numpy==2.2.4
streamlit==1.44.0
plotly==6.0.1
sqlalchemy==2.0.40
mysql-connector-python==9.2.0
openai==1.69.0
python-dotenv==1.1.0
